dev_requirements.txt
aioopenssl==0.5.1
aiosasl==0.4.1
aioxmpp==0.11.0
altgraph==0.17
ansicolor==0.2.6
ansicolors==1.1.8
atpublic==2.1.1
Babel==2.8.0
cffi==1.14.3
click==7.1.2
cryptography==3.2.1
Cython==0.29.21
dnspython==1.16.0
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.6.1
markdown2==2.3.10
MarkupSafe==1.1.1
multidict==4.7.6
pefile==2019.4.18
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.9
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.11
pyOpenSSL==19.1.0
pytz==2020.4
pywin32-ctypes==0.2.0
six==1.15.0
sortedcollections==1.2.1
sortedcontainers==2.3.0
SQLAlchemy==1.3.20
tzlocal==1.5.1
waitress==1.4.4
Werkzeug==1.0.1
